earnings_volatility_yfinance/requirements.txt
# Earnings Volatility Trading Bot Dependencies (yfinance version)
# Note: Python 3.11+ is required (specified in Dockerfile)

# Data source
yfinance>=0.2.0

# API clients
alpaca-py>=0.30.0
supabase>=2.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Retry logic
tenacity>=8.2.0

# Utilities
pytz>=2023.3
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0  # For earnings calendar API

